Extract base class for WCF tests
